fix: servers API returning `port` as a string -> integer

+++ app/Models/Server.php
@@ -813,7 +813,7 @@ $schema://$host {
     {
         return Attribute::make(
             get: function ($value) {
-                return preg_replace('/[^0-9]/', '', $value);
+                return (int) preg_replace('/[^0-9]/', '', $value);
             }
         );
     }
